feature(source/startAnimationDevelopment): initialSaga cleanup
